tests/wpt/web-platform-tests/webmessaging/with-options/one-arg.html

<!doctype html>
<title>just one argument</title>
<script src="/resources/testharness.js"></script>
<script src="/resources/testharnessreport.js"></script>
<div id="log"></div>
<script>
test(function() {
  postMessage('');
  onmessage = this.step_func(function(e) {
    assert_equals(e.data, '');
    this.done();
  });
});
</script>
